Dont remake raw experiment

+++ Determine_prefiltering_thresholds.R
@@ -41,12 +41,16 @@ sample_IDs <- read.csv(paste0(location, "sarah_projects/SCMDA231mets_chrcha/samp
 
 # Load all samples and save before editing
 data_directory <- paste0(location, "sarah_projects/SCMDA231mets_chrcha/raw_data/data")
-raw_experiment <- read10xCounts(paste0(list.files(data_directory, full.names = TRUE), '/outs/filtered_feature_bc_matrix'), col.names = TRUE, sample.names = list.files(data_directory, full.names = FALSE))
 
-if (place == "wolfpack") {
-  saveRDS(raw_experiment, "all_data/Raw_experiment_all_samples_old_labels.rds")
-} else {
-  print("Working local")
+if(file.exists("all_data/Raw_experiment_all_samples_old_labels.rds")) {
+  raw_experiment <- readRDS("all_data/Raw_experiment_all_samples_old_labels.rds")
+  } else {
+  raw_experiment <- read10xCounts(paste0(list.files(data_directory, full.names = TRUE), '/outs/filtered_feature_bc_matrix'), col.names = TRUE, sample.names = list.files(data_directory, full.names = FALSE))
+  if (place == "wolfpack") {
+    saveRDS(raw_experiment, "all_data/Raw_experiment_all_samples_old_labels.rds")
+  } else {
+    print("Working local")
+  }
 }
 
 # Rename colData Sample slot. Add Tissue, Met, Replicate information.
@@ -93,18 +97,17 @@ undetectedGenes <- rowSums(counts(raw_experiment)) == 0
 raw_experiment <- raw_experiment[!undetectedGenes, ] # Remove undetected genes
 
 # Calculate SAVER standard error estimates and plot relative to library size. Likely observed two distinct populations, one with low library size and high standard error, the other with higher library size and lower standard error (set filter parameters to capture this population)
-row_names <- split(rownames(filtered_exp), ceiling(seq_along(rownames(filtered_exp))/2500))
+row_names <- split(rownames(raw_experiment), ceiling(seq_along(rownames(raw_experiment))/2500))
 names(row_names)
 
 for(i in names(row_names)){
-  saver.new <- saver(counts(filtered_exp), pred.genes = which(rownames(filtered_exp) %in% row_names[[i]]), pred.genes.only = TRUE, do.fast = FALSE, ncores = 32)
+  saver.new <- saver(counts(raw_experiment), pred.genes = which(rownames(raw_experiment) %in% row_names[[i]]), pred.genes.only = TRUE, do.fast = FALSE, ncores = 32)
   if(exists(saver.out)){
     saver.out <- combine.saver(list(saver.out, saver.new))
   } else {
     saver.out <- saver.new
     }
   } # Split genes into multiple groups of 2500 to avaoid memory issues and then recombine into single object
-
 saveRDS(saver.out, "SAVER_run_raw_experiment.rds")
 
 pdf("Unfiltered_SAVER_SEvsLibSize.pdf")
